- Restructuring project directory and files
