+++ GUI/Qt/Models/passwordstrengthchecker.cpp
