+++ GUI/Qt/Models/passwordstrengthchecker.h
