+++ GUI/Qt/Models/randomseedgenerator.cpp
