+++ GUI/Qt/Models/randomseedgenerator.h
