+++ GUI/Qt/qml/ContainerEncrypted.qml
