+++ GUI/Qt/qml/MessageBar.qml
